Improve error handling and functionality of vimux

+++ zsh/.zshrc
@@ -16,7 +16,7 @@ setopt HIST_IGNORE_SPACE
 
 # Environment
 
-NVIM_SESSION_DIR=$HOME/.local/share/nvim/sessions
+NVIM_SERVERS_DIR=$HOME/.local/share/nvim/servers
 
 export TERM=xterm-256color
 export VISUAL=vi
@@ -52,46 +52,67 @@ _comp_options+=(globdots)
 # Utilities
 
 if command -v nvim >/dev/null 2>&1; then
-    vi() {
-        if [[ -n $1 ]]; then
-            nvim "$@"
-            return $?
-        fi
-        session_name=$(basename $PWD)
-        if [[ -S $NVIM_SESSION_DIR/$session_name ]]; then
-            echo -e "\033[31mSession $session_name already exists\033[0m" >&2
-            echo "Run 'vimux a[ttach] $session_name'" >&2
+    _vi_server_exists() {
+        if [[ -z $1 ]]; then
+            setopt localoptions null_glob
+            local servers=("$NVIM_SERVERS_DIR"/*)
+            return $(( ${#servers[@]} == 0 ))
+        elif [[ -S $NVIM_SERVERS_DIR/$1 ]]; then
+            return 0
+        else
             return 1
         fi
-        mkdir -p "$NVIM_SESSION_DIR"
-        nvim --listen "$NVIM_SESSION_DIR/$session_name"
     }
 
     vimux() {
         case "$1" in
         a | attach)
-            session_name=$2
-            if [[ -z $session_name ]]; then
-                session_name=$(ls "$NVIM_SESSION_DIR" | fzf) 
+            if [[ -n $2 ]] && ! _vi_server_exists "$2"; then
+                echo "\033[31mServer $2 doesn't exist\033[0m" >&2
+                echo "'vimux ls' will show the available servers" >&2
+                return 1
             fi
-            server=$NVIM_SESSION_DIR/$session_name
+            local server_name=$2
+            if [[ -z $server_name ]]; then
+                server_name=$(ls "$NVIM_SERVERS_DIR" | fzf)
+            fi
+            local server=$NVIM_SERVERS_DIR/$server_name
             pkill -f "nvim .*--server $server --remote-ui"
             nvim --server "$server" --remote-ui
             ;;
         ls | list)
-            format='%-8s %-20s %-30s %7s\n'
-            printf "$format" PID NAME CWD BUFFERS
-            for server in "$NVIM_SESSION_DIR"/*; do
-                name=$(basename "$server")
-                pid=$(nvim --headless --server "$server" --remote-expr 'getpid()')
-                cwd=$(nvim --headless --server "$server" --remote-expr 'getcwd()')
-                buffers=$(nvim --headless --server "$server" --remote-expr 'len(getbufinfo({"buflisted": 1}))')
-                printf "$format" "$pid" "$name" "$cwd" "$buffers"
+            if ! _vi_server_exists; then
+                echo "No servers are currently running"
+                return 0
+            fi
+            format='%-8s %-20s %-30s\n'
+            printf "$format" PID NAME CWD
+            for server in "$NVIM_SERVERS_DIR"/*; do
+                local name=$(basename "$server")
+                local pid=$(nvim --headless --server "$server" --remote-expr 'getpid()')
+                local cwd=$(nvim --headless --server "$server" --remote-expr 'getcwd()')
+                printf "$format" "$pid" "$name" "$cwd"
             done
             ;;
         *)
             echo "Usage: vimux (a|attach)|(ls|list)" >&2
             return 1
         esac
+    }
+
+    vi() {
+        if [[ -n $1 ]]; then
+            nvim "$@"
+            return $?
+        fi
+        local server_name=$(basename $PWD)
+        if _vi_server_exists "$server_name"; then
+            echo "Server $server_name is already running"
+            echo "Connecting to existing neovim instance"
+            vimux attach "$server_name"
+            return "$?"
+        fi
+        mkdir -p "$NVIM_SERVERS_DIR"
+        nvim --listen "$NVIM_SERVERS_DIR/$server_name"
     }
 fi
